ref: allocate key_fetcher to seperate file
